Fix ms3_copy for non-alphanumeric characters
The S3 API specification states that x-amz-copy-source should be
urlencoded. libmarias3 wasn't doing this but Amazon's S3 implementation
appeared to be forgiving enough that it didn't matter. When using
compatible APIs such as MinIO this requirement became stricter.

The code now urlencodes the source path for ms3_copy / ms3_move.

In addition this patch adds support for MinIO in the test suite by
adding environment variables for port and disabling https.