Add custom callback mechanism (#113)
* Allow custom read callbacks to be used

This allows handling of data in smaller chunks if the ms3_get result is
too large to fit into memory.

* Use the configured buffer size in curl

This way the chunks that curl will return will be the same size that the
internal buffer will be. For the custom read callback, this will allow
the data to be handled in smaller chunks.

* Fix copyright year in read_cb.c

The year is 2023.