Allow for disabling content type
Curl automatically sets the content type of
`application/x-www-form-urlencoded`, whilst this hasn't caused us any
problems in the past, Huawei's S3 implementation does not like it. The
two possible options are to remove the content type header, or use mime
multipart uploads.

Multipart uploads could be implemented eventually, but are only really
needed for very large files (> 4GB). So, this patch adds an option to do
the former, disabling the content type. This is not a default option at
the moment and needs to be set using:

```
ms3_set_option(ms3, MS3_OPT_NO_CONTENT_TYPE, NULL);
```